move IWantAmountOrPercentTypeXml to constants
