--- src/main/java/edu/cornell/kfs/module/purap/businessobject/xml/IWantAmountOrPercentTypeXml.java
@@ -1,20 +1,0 @@
-package edu.cornell.kfs.module.purap.businessobject.xml;
-
-import jakarta.xml.bind.annotation.XmlEnum;
-import jakarta.xml.bind.annotation.XmlType;
-
-@XmlType(name = "amountOrPercentType", namespace = "http://www.kuali.org/kfs/purap/types")
-@XmlEnum
-public enum IWantAmountOrPercentTypeXml {
-
-    A, P;
-
-    public String value() {
-        return name();
-    }
-
-    public static IWantAmountOrPercentTypeXml fromValue(String v) {
-        return valueOf(v);
-    }
-
-}
